Created new file Factory
Factory file incorperates an interface for creating objects in a super-
class, allowing subclasses to alter the type of object that will be
created.

Main objects used in conjunction with the Factory method is Cookie and
Ghost and have multiple instances of the same class.

Adding more object of Ghosts and Cookies would require changes to the
entire codebase. Avoid using unnecessary changes to the codebase.