Imported JavaFx library
JavaFx library is responsible for the graphics and animation of the
user interface. Judging the complexity of adding the library into the
library to the java project, final additions of the library
compile correctly and our recognised within the java project.

JavaFx imports persistant across the entire java project. Main
compilation of the Main class works, further running causing runtime
error with JavaFx library.