Refactor Manager class with optimised method
Most of the methods within the class Manager were either wrongly defined
being that the names of the methods were incorrect or the whether the
methods were to be public or private methods were consistent.

Manager class contains many methods, some of which do not belong within
the class and must be delegated to other clases.

Moudle.info file added to commit branch, later use of moudles to be
further added.